+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 2025.12.14 (2025-12-08)
 
 - Eliminate CDPD dictionary format with pure ProfileConfig implementation
+- Refactor CCU backup
 
 # Version 2025.12.13 (2025-12-07)